fix:[result]오류수정 및 폴더 생성(type폴더, util/rank 폴더)
