--- src/util/supabase/server-action.ts
@@ -1,17 +1,0 @@
-import { createClient } from './server';
-
-// 현재 사용자 조회
-export const fetchCurrentUser = async () => {
-  const supabase = await createClient();
-  const {
-    data: { user },
-    error,
-  } = await supabase.auth.getUser();
-
-  if (error || !user) {
-    console.error(error);
-    return null;
-  }
-
-  return user;
-};
